nambah acces program

+++ app/Http/Controllers/AccessProgramController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers;
 use App\Models\AccessProgram;
 use App\Http\Controllers\Controller;
 use App\Models\AccessMenu;
+use App\Models\DataWarga;
 use App\Models\Menu;
 use App\Models\Program;
 use App\Models\Role;
 use App\Models\SubMenu;
+use App\Models\UpdateKerja;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
@@ -38,18 +40,30 @@ class AccessProgramController extends Controller
      */
     public function store(Request $request)
     {
+        $user = User::find($request->user_id);
+        $data_warga = DataWarga::find($user->data_warga_id);
+        if ($data_warga->status == "Bekerja") {
 
-        $data = new AccessProgram();
-        $data->user_id = $request->user_id;
-        $data->program_id = $request->program_id;
+            $data = new AccessProgram();
+            $data->user_id = $request->user_id;
+            $data->program_id = $request->program_id;
 
-        $cek_access = AccessProgram::where('user_id', $request->user_id)->where('program_id', $request->program_id);
-        if ($cek_access->count() < 1) {
-            $data->save();
-            return redirect()->back()->with('sukses', 'Wahhhhh Mantappp luar biasa atos bergabung kana program');
+            $cek_access = AccessProgram::where('user_id', $request->user_id)->where('program_id', $request->program_id);
+            if ($cek_access->count() < 1) {
+                $data->save();
+                return redirect()->back()->with('sukses', 'Wahhhhh Mantappp luar biasa atos bergabung kana program');
+            } else {
+                $cek_access->forceDelete();
+                return redirect()->back()->with('kuning', 'Yahhhhhh Akses kana program atos di putus');
+            }
+
+            $update = new UpdateKerja();
+            $update->user_id = $request->user_id;
+            $update->status = $request->status;
+
+            $update->save();
         } else {
-            $cek_access->forceDelete();
-            return redirect()->back()->with('kuning', 'Yahhhhhh Akses kana program atos di putus');
+            return redirect()->back()->with('kuning', 'Duhhhh Punten teu acan tiasa ngiringan program ieu, cek heula status peerjaan na. Program ieu tiasa di ikuti ku anggota nu tos damel');
         }
     }
